% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_game_states.R
\name{are_identical_arenas}
\alias{are_identical_arenas}
\title{Are identical arena's}
\usage{
are_identical_arenas(arena_1, arena_2)
}
\arguments{
\item{arena_1}{First arena.}

\item{arena_2}{Second arena.}
}
\value{
Logical indicating whether arena's are identical (\code{TRUE} ir \code{FALSE}).
}
\description{
This function determines whether two arena's are identical.
}
