% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_game_states.R
\name{are_identical_game_states}
\alias{are_identical_game_states}
\title{Are identical game state}
\usage{
are_identical_game_states(game_state_1, game_state_2)
}
\arguments{
\item{game_state_1}{First game state.}

\item{game_state_2}{Second game state.}
}
\value{
Logical indicating whether game states are identical (\code{TRUE} or
\code{FALSE}).
}
\description{
This function determines whether two game states are identical.
}
