% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_game_states.R
\name{are_identical_ghosts}
\alias{are_identical_ghosts}
\title{Are identical lists of ghosts}
\usage{
are_identical_ghosts(ghosts_1, ghosts_2)
}
\arguments{
\item{ghosts_1}{First list with ghosts.}

\item{ghosts_2}{Second list with ghosts.}
}
\value{
Logical indicating whether lists of ghosts are identical (\code{TRUE} or
\code{FALSE})
}
\description{
This function determines whether two lists of ghosts are identical.
}
