% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_game_states.R
\name{are_identical_players}
\alias{are_identical_players}
\title{Are identical players.}
\usage{
are_identical_players(player_1, player_2)
}
\arguments{
\item{player_1}{First player.}

\item{player_2}{Second player.}
}
\value{
Logical indicating whether players are identical (\code{TRUE} or \code{FALSE}).
}
\description{
This function determines whether two players are identical.
}
