% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arena.R
\name{Arena}
\alias{Arena}
\title{R6 Class representing an arena}
\description{
R6 Class representing an arena

R6 Class representing an arena
}
\details{
An arena has a layout.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Arena$new()}}
\item \href{#method-print}{\code{Arena$print()}}
\item \href{#method-clone}{\code{Arena$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new arena object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arena$new(layout)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{layout}}{Dataframe with the layout of the arena.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Arena} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arena$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Arena$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
