% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_action.R
\name{check_action}
\alias{check_action}
\title{Check if the action is valid}
\usage{
check_action(action)
}
\arguments{
\item{action}{a list representing an action
Will \link{stop} if not}
}
\description{
Check if the action is valid
}
