% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arena.R
\name{check_arena}
\alias{check_arena}
\title{Check if the arena is valid
Needs to have at least 3 lines
All lines need to have the same amount of characters}
\usage{
check_arena(arena_text)
}
\arguments{
\item{arena_text}{Blablabla}
}
\description{
Will \link{stop} if not
}
