% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_game_state.R
\name{check_game_state}
\alias{check_game_state}
\title{Verify whether the game state is valid, and stop the game when it isn't.}
\usage{
check_game_state(game_state)
}
\description{
Verify whether the game state is valid, and stop the game when it isn't.
}
