% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_action.R
\name{create_action}
\alias{create_action}
\title{Create an action for the player to go left}
\usage{
create_action(action_type)
}
\arguments{
\item{action_type}{the type of action}
}
\description{
Create an action for the player to go left
}
