% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_action.R
\name{create_action_player_left}
\alias{create_action_player_left}
\title{Create an action for the player to go left}
\usage{
create_action_player_left()
}
\description{
Create an action for the player to go left
}
