% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_action.R
\name{create_action_player_right}
\alias{create_action_player_right}
\title{Action to make the player go right}
\usage{
create_action_player_right()
}
\description{
Action to make the player go right
}
\note{
this function must call \link{create_action} to
inherit everything what an action is
}
