% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game_state.R
\name{create_game_state}
\alias{create_game_state}
\title{Create a game state.}
\usage{
create_game_state(arena, player)
}
\arguments{
\item{arena}{An arena.}

\item{player}{A player.}
}
\value{
A game state.
}
\description{
Create a game state.
}
