% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_arena_table.R
\name{create_html_arena_table}
\alias{create_html_arena_table}
\title{Create the HTML of the arena as a table}
\usage{
create_html_arena_table(game_state)
}
\value{
the table as one line of HTML text
}
\description{
Create the HTML of the arena as a table
}
