% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_player.R
\name{create_player}
\alias{create_player}
\title{Create a player}
\usage{
create_player(x = 1, y = 1, size = 1)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{size}{size}
}
\value{

}
\description{
Create a player
}
