% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_ghosts.R
\name{create_test_ghosts}
\alias{create_test_ghosts}
\title{Create a list with ghosts for testing purposes}
\usage{
create_test_ghosts()
}
\description{
Create a list with ghosts for testing purposes
}
