% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ui.R
\name{create_ui}
\alias{create_ui}
\title{Visualizes the state of the game in HTML}
\usage{
create_ui(game_state)
}
\arguments{
\item{game_state}{a game state,
as created by \link{create_game_state}}
}
\value{
a \code{c("html", "character")}
}
\description{
Visualizes the state of the game in HTML
}
