% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_action.R
\name{do_action}
\alias{do_action}
\title{Do action}
\usage{
do_action(game_state, action)
}
\arguments{
\item{game_state}{Game state.}

\item{action}{Action.}
}
\value{
New game state.
}
\description{
Apply action to player.
}
