% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Game_State.R
\name{Game_State}
\alias{Game_State}
\title{R6 Class representing a game state}
\description{
R6 Class representing a game state

R6 Class representing a game state
}
\details{
A game state has a player and an arena
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Game_State$new()}}
\item \href{#method-print}{\code{Game_State$print()}}
\item \href{#method-clone}{\code{Game_State$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new game state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game_State$new(player, arena)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player}}{Player.}

\item{\code{arena}}{Arena.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Game_State} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game_State$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Game_State$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
