% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameState.R
\name{GameState}
\alias{GameState}
\title{R6 Class representing a game state}
\description{
R6 Class representing a game state

R6 Class representing a game state
}
\details{
A game state has a player, list of ghosts and an arena.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameState$new()}}
\item \href{#method-print}{\code{GameState$print()}}
\item \href{#method-do_action}{\code{GameState$do_action()}}
\item \href{#method-clone}{\code{GameState$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new game state object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$new(player, ghosts, arena)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{player}}{Player.}

\item{\code{ghosts}}{List of ghosts.}

\item{\code{arena}}{Arena.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{GameState} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$print(verbose = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_action"></a>}}
\if{latex}{\out{\hypertarget{method-do_action}{}}}
\subsection{Method \code{do_action()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$do_action(action)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameState$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
