% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_moves.R
\name{get_best_moves}
\alias{get_best_moves}
\title{Get best moves}
\usage{
get_best_moves(player, arena)
}
\arguments{
\item{player}{Player}

\item{arena}{Arena.}
}
\value{
Named character vector with best moves for every position in the
arena.
}
\description{
This function uses reinforcement learning to determine the best moves
for a ghost for every position in the arena, taking into account the
position of the player.
}
