% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ghost.R
\name{Ghost}
\alias{Ghost}
\title{R6 Class representing a ghost}
\description{
R6 Class representing a ghost

R6 Class representing a ghost
}
\details{
A ghost has an x and y coordinate,
and a personality (chase, ambush,
freak or idiot).
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ghost$new()}}
\item \href{#method-print}{\code{Ghost$print()}}
\item \href{#method-clone}{\code{Ghost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new ghost object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ghost$new(x, y, personality)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x coordinate.}

\item{\code{y}}{y coordinate.}

\item{\code{personality}}{Personality (chase, ambush,
freak or idiot).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Ghost} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ghost$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ghost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
