% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_game_state.R
\name{is_game_state}
\alias{is_game_state}
\title{Check if an object is a valid game state.}
\usage{
is_game_state(game_state)
}
\arguments{
\item{game_state}{Game state.}
}
\value{
logical, \code{TRUE} for a valid and \code{FALSE} for an invalid game state.
}
\description{
Check if an object is a valid game state.
}
