% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_player.R
\name{is_player}
\alias{is_player}
\title{Check if an object is a valid player.}
\usage{
is_player(player)
}
\arguments{
\item{player}{Player.}
}
\value{
logical, \code{TRUE} for a valid and \code{FALSE} for an invalid player.
}
\description{
Check if an object is a valid player.
}
