% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Action.R
\name{Left}
\alias{Left}
\title{R6 Class representing a "go left" action}
\description{
R6 Class representing a "go left" action

R6 Class representing a "go left" action
}
\section{Super class}{
\code{\link[snackman:Action]{snackman::Action}} -> \code{Left}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Left$new()}}
\item \href{#method-clone}{\code{Left$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="snackman" data-topic="Action" data-id="print">}\href{../../snackman/html/Action.html#method-print}{\code{snackman::Action$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new go left object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Left$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{Left} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Left$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
