% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ai_model.R
\docType{data}
\name{model}
\alias{model}
\title{AI model for predicting the best moves for ghosts.}
\format{
An object of class \code{rl} of length 8.
}
\usage{
model
}
\description{
The model was trained using the function \code{\link{train_ai_model}}.
}
\keyword{datasets}
