% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\name{Player}
\alias{Player}
\title{R6 Class representing a player}
\description{
R6 Class representing a player

R6 Class representing a player
}
\details{
A player has an x and y coordinate,
size and name.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Player$new()}}
\item \href{#method-print}{\code{Player$print()}}
\item \href{#method-clone}{\code{Player$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new player object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$new(x, y, size, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x coordinate.}

\item{\code{y}}{y coordinate.}

\item{\code{size}}{Size (1, 2 or 3).}

\item{\code{name}}{Name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Player} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
