% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_best_move.R
\name{predict_best_move}
\alias{predict_best_move}
\title{Predict best move for ghosts}
\usage{
predict_best_move(game_state)
}
\arguments{
\item{game_state}{Game state.}
}
\value{
Character vector containing the best move for each ghost in the game.
}
\description{
This function predicts the best moves for the ghosts in a game.
}
