% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_game_state.R
\name{print_game_state}
\alias{print_game_state}
\title{Show the game state in a pretty way}
\usage{
print_game_state(game_state)
}
\arguments{
\item{game_state}{a game state, as created by \link{create_game_state}}
}
\description{
Show the game state in a pretty way
}
