% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackman_loader.R
\name{snackman_html}
\alias{snackman_html}
\title{Snackman game HTML}
\usage{
snackman_html(header = "Snackman")
}
\arguments{
\item{header}{The title to display over the game.}
}
\value{
Panel that will load the snackman game.
}
\description{
fun game inspired on pacman
}
\examples{
snackman_html()
snackman_html(header = "Custom header")
}
