% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snackman_loader.R
\name{snackman_loader}
\alias{snackman_loader}
\title{Snackman game HTML}
\usage{
snackman_loader(header = "Snackman")
}
\arguments{
\item{header}{The title to display over the game screen.}
}
\value{
Panel that will load the snackman game.
}
\description{
fun game inspired on pacman
}
\examples{
\dontrun{
library(shiny)
library(snackman)

# Add here exmaple

}

}
