% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ai_model.R
\name{train_ai_model}
\alias{train_ai_model}
\title{Train AI model}
\usage{
train_ai_model(player, arena)
}
\arguments{
\item{player}{Player}

\item{arena}{Arena.}
}
\value{
Named character vector with best moves for every position in the
arena.
}
\description{
This function uses reinforcement learning to train a model to predict the
best move for a ghost, taking into account the position of the player.
}
