% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_game_state.R
\name{check_game_state}
\alias{check_game_state}
\title{Verify whether the game state is valid, and stop the game when it isn't.}
\usage{
check_game_state(game_state)
}
\arguments{
\item{game_state}{Game state.}
}
\description{
Verify whether the game state is valid, and stop the game when it isn't.
}
\examples{
arena <- create_arena()
player <- list()
player$x <- 100.0
player$y <- 200.0
player$size <- 1
game_state <- create_game_state(arena, player)
check_game_state(game_state)
}
