% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_action.R
\name{create_action_player_left}
\alias{create_action_player_left}
\title{Action to make the player go left}
\usage{
create_action_player_left()
}
\description{
Action to make the player go left
}
\note{
this function must call \link{create_action} to
inherit everything what an action is
}
