% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_action.R
\name{create_action_player_right}
\alias{create_action_player_right}
\title{Create an action for the player to go right}
\usage{
create_action_player_right()
}
\description{
Create an action for the player to go right
}
