% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game_state.R
\name{create_game_state}
\alias{create_game_state}
\title{Create a game state.}
\usage{
create_game_state(arena, player)
}
\arguments{
\item{a}{An arena.}

\item{p}{A player.}
}
\value{
A game state.
}
\description{
Create a game state.
}
\examples{
arena <- create_arena()
player <- list()
player$x <- 100.0
player$y <- 200.0
player$size <- 1
game_state <- create_game_state(arena, player)
}
