% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_best_moves.R
\name{get_best_moves}
\alias{get_best_moves}
\title{Get best moves}
\usage{
get_best_moves(x_pos, y_pos, arena)
}
\arguments{
\item{x_pos}{x position of the enemy in the arena.}

\item{y_pos}{y position of the enemy in the arena.}

\item{arena}{Arena.}
}
\value{
Named character vector with best moves for every position in the
arena, taking into account the position of the enemy.
}
\description{
This function uses reinforcement learning to determine the best moves
for a player for every position in the arena, taking into account the
position of the enemy.
}
\details{
See https://cran.r-project.org/web/packages/ReinforcementLearning/
vignettes/ReinforcementLearning.html
}
