% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_game_state.R
\name{is_game_state}
\alias{is_game_state}
\title{Check if an object is a valid game state.}
\usage{
is_game_state(game_state)
}
\arguments{
\item{game_state}{Game state.}
}
\value{
logical, \code{TRUE} for a valid and \code{FALSE} for an invalid game state.
}
\description{
Check if an object is a valid game state.
}
\examples{
arena <- create_arena()
player <- list()
player$x <- 100.0
player$y <- 200.0
player$size <- 1
game_state <- create_game_state()
is_game_state(game_state)
}
