% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_ai_model.R
\name{train_ai_model}
\alias{train_ai_model}
\title{Train AI model}
\usage{
train_ai_model(player, arena)
}
\arguments{
\item{player}{Player}

\item{arena}{Arena.}
}
\value{
Reinforcement learning model predicting the best next move.
}
\description{
This function uses reinforcement learning to train a model to predict the
best move for a ghost, taking into account the position of the player.
}
\details{
The trained model is saved as a \code{.rda} file in the \code{data} map:
\code{usethis::use_data(ai_model, internal = TRUE, overwrite = TRUE)}
}
\examples{
player <- create_test_player()
arena <- create_test_arena()
ai_model <- train_ai_model(player, arena)
# Display the best possible action in every state
ReinforcementLearning::computePolicy(ai_model)
# Write state-action table to the screen
print(ai_model)
# Additional diagnostics regarding the model
summary(ai_model)
# Apply the model to predict best move
predict(ai_model, "3 4")
}
